% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{addcovar}
\alias{addcovar}
\title{Additive covariates to use in the mapping model.}
\format{A matrix with 500 rows and 5 columns:
\describe{
  \item{sexM}{numeric indicator for sex (0 = female, 1 = male)}
  \item{DOwave2}{numeric indicator for DO wave 2}
  \item{DOwave3}{numeric indicator for DO wave 3}
  \item{DOwave4}{numeric indicator for DO wave 4}
  \item{DOwave5}{numeric indicator for DO wave 5}
}}
\usage{
addcovar
}
\description{
Additive covariates to use in the mapping model.
}
\keyword{datasets}
