% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr}
\alias{expr}
\title{Gene expression data for genes on Chr 13.}
\format{A matrix with 378 rows and 800 columns:
\describe{
   Samples are in rows and genes are in columns.
}}
\usage{
expr
}
\description{
Gene expression data for genes on Chr 13.
}
\keyword{datasets}
