% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genoprobs}
\alias{genoprobs}
\title{Founder allele probabilities for each mouse on Chr 13.}
\format{A list with 20 elements:
\describe{
   500 rows by 8 founders by 1000 markers.
}}
\usage{
genoprobs
}
\description{
All chromosomes except for Chr 13 are st to NA to save space.
Also, the markers have been down-sampled to 1000 markers.
The element for Chr 13 contains a three-dimensional array of 
allele probabilities with samples in rows, founders in columns
and markers in slices.
}
\keyword{datasets}
