% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_probs.R
\name{get_snp_probs}
\alias{get_snp_probs}
\title{Get the SNP probs nearest to the given chr and position.}
\usage{
get_snp_probs(chr, pos, genoprobs, map, query_fxn)
}
\value{
an n x 1 numeric matrix.
}
\description{
chr: character indicating the chromosome to use. This must match
     one of the names of genoprobs and map.
pos: numeric value in Mb indicating the position on the chromosome.
genoprobs: list containing 3-dimensional arrays of allele probs, one 
           for each chromosome. In qtl2 format.
           Samples in rows, 8 founders in columns, markers in slices.
map: list containing numeric vectors of marker positions in Mb.
query_fxn: function for querying the CC_SNP data base. Obtain this
}
\examples{
query_fxn = create_variant_query_func(cc_dbfile, filter = "type=='snp'")
get_snp_probs = function(chr = "1", pos = 5.0, genoprobs = genoprobs, map = map, query_fxn = query_fxn)
}
