% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map}
\alias{map}
\title{Physical locations for the markers.}
\format{A list with 20 elements:
\describe{
   \item{1}{Marker locations on Chr 1}
   \item{1}{Marker locations on Chr 2}
   \item{1}{Marker locations on Chr 3}
   \item{1}{Marker locations on Chr 4}
   \item{1}{Marker locations on Chr 5}
   \item{1}{Marker locations on Chr 6}
   \item{1}{Marker locations on Chr 7}
   \item{1}{Marker locations on Chr 8}
   \item{1}{Marker locations on Chr 9}
   \item{1}{Marker locations on Chr 10}
   \item{1}{Marker locations on Chr 11}
   \item{1}{Marker locations on Chr 12}
   \item{1}{Marker locations on Chr 13}
   \item{1}{Marker locations on Chr 14}
   \item{1}{Marker locations on Chr 15}
   \item{1}{Marker locations on Chr 16}
   \item{1}{Marker locations on Chr 17}
   \item{1}{Marker locations on Chr 18}
   \item{1}{Marker locations on Chr 19}
   \item{1}{Marker locations on Chr X}
}}
\usage{
map
}
\description{
Each element is a vector of marker locations in Mb.
}
\keyword{datasets}
