% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate.R
\name{mediate}
\alias{mediate}
\title{Regress one phenotype on the given genes at a single marker.}
\usage{
mediate(pheno, chr, pos, type = c("haplo", "snp"), expr, ensembl, genoprobs,
  addcovar, cc_dbfile)
}
\arguments{
\item{pheno:}{numeric matrix with one column. Sample IDs must be in rownames and must match rownames in genoprobs.}

\item{chr:}{character containing a single chromosome ID.}

\item{pos:}{numeric value indicating the position in Mb at which to mediate.}

\item{type:}{character value that is either 'haplo' or 'snp', indicating which type of genotype values to mediate upon.}

\item{expr:}{numeric matrix containing the gene expression values. Sample names in rownames and Ensembl IDs in colnames.}

\item{ensembl:}{data.frame containing gene annotation as obtained from \code{get_ensembl_gene}.}

\item{genoprobs:}{list containing 3-dimensional arrays of allele probs, one for each chromosome. In qtl2 format. Samples in rows, 8 founders in columns, markers in slices.}

\item{addcovar:}{numeric matrix of additive covariates without the intercept. Sample IDs must be in rownames. Typically creates using model.matrix(~covar1 + covar2, data = data)[,-1].}

\item{cc_dbfile:}{full path to the CC_SNP database file created by Karl Broman.}
}
\value{
data.frame containing the mediation results.
}
\description{
Regress one phenotype on the given genes at a single marker.
}
