% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Phenotype values for one phenotype.}
\format{A data.frame with 378 rows and 1 column:
\describe{
   \item{phenotype}{A numeric phenotype.}
}}
\usage{
pheno
}
\description{
Phenotype values for one phenotype.
}
\keyword{datasets}
