% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_med.R
\name{plot_med}
\alias{plot_med}
\title{Plot mediation results.}
\usage{
plot_med(med, threshold = -1)
}
\arguments{
\item{med}{data.frame (or tibble) containing the mediation results returned by \code{\link{mediate}}.}

\item{threshold}{floating point number indicating the number of standard deviations of the LOD above or below which genes will be labelled. Negative numbers suppress labels. Default = -1.}
}
\value{
a mediation plot.
}
\description{
Plot mediation results.
}
