\name{qtl2med-package}
\alias{qtl2med-package}
\alias{qtl2med}
\docType{package}
\title{
\packageTitle{qtl2med}
}
\description{
\packageDescription{qtl2med}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qtl2med}
\packageIndices{qtl2med}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{qtl2med}

Maintainer: \packageMaintainer{qtl2med}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
